+++ vm/devices/storage/nvme_test/src/pci.rs
@@ -193,11 +193,11 @@ impl NvmeFaultController {
     }
 
     /// Reads from the virtual BAR 0.
-    pub fn read_bar0(&mut self, addr: u16, data: &mut [u8]) -> IoResult {
+    pub fn read_bar0(&mut self, addr: u64, data: &mut [u8]) -> IoResult {
         if data.len() < 4 {
             return IoResult::Err(IoError::InvalidAccessSize);
         }
-        if addr & (data.len() - 1) as u16 != 0 {
+        if addr & (data.len() as u64 - 1) != 0 {
             return IoResult::Err(IoError::UnalignedAccess);
         }
 
@@ -251,7 +251,7 @@ impl NvmeFaultController {
     }
 
     /// Writes to the virtual BAR 0.
-    pub fn write_bar0(&mut self, addr: u16, data: &[u8]) -> IoResult {
+    pub fn write_bar0(&mut self, addr: u64, data: &[u8]) -> IoResult {
         if addr >= 0x1000 {
             // Doorbell write.
             let base = addr - 0x1000;
@@ -263,14 +263,14 @@ impl NvmeFaultController {
                 return IoResult::Err(IoError::InvalidAccessSize);
             };
             let value = u32::from_ne_bytes(data);
-            self.workers.doorbell(db_id, value);
+            self.workers.doorbell(db_id as u16, value);
             return IoResult::Ok;
         }
 
         if data.len() < 4 {
             return IoResult::Err(IoError::InvalidAccessSize);
         }
-        if addr & (data.len() - 1) as u16 != 0 {
+        if addr & (data.len() as u64 - 1) != 0 {
             return IoResult::Err(IoError::UnalignedAccess);
         }
