+++ vm/devices/virtio/virtio/src/transport/mmio.rs
@@ -164,7 +164,7 @@ impl VirtioMmioDevice {
 
 impl VirtioMmioDevice {
     pub(crate) fn read_u32(&mut self, address: u64) -> u32 {
-        let offset = (address & 0xfff) as u16;
+        let offset = address & 0xfff;
         assert!(offset & 3 == 0);
         match VirtioMmioRegister(offset) {
             VirtioMmioRegister::MAGIC_VALUE => u32::from_le_bytes(*b"virt"),
@@ -272,7 +272,7 @@ impl VirtioMmioDevice {
     }
 
     pub(crate) fn write_u32(&mut self, address: u64, val: u32) {
-        let offset = (address & 0xfff) as u16;
+        let offset = address & 0xfff;
         assert!(offset & 3 == 0);
         let queue_select = self.queue_select as usize;
         let queues_locked = self.device_status.driver_ok();
